LayoutTests/accessibility/progressbar-expected.txt

X X
This test makes sure that progress element can be accessed by Accessibility FW.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS obj.intValue is 7
PASS obj.intValue is 5
PASS obj.intValue is 7
PASS obj.role is 'AXRole: AXProgressIndicator'
PASS obj.intValue is 0
PASS obj.role is 'AXRole: AXProgressIndicator'
PASS successfullyParsed is true

TEST COMPLETE
